refactor(goals): goal vocabulary moves to lib/classes
The agent entity union imports payload models only from lib/classes (the
day_plan precedent — 'classes is the shared vocabulary that lets
features/agents not depend on features'). PR 1's goalSpecVersion and
goalProgress variants embed GoalCriterion and GoalTrackStatus, so the
vocabulary (enums, window, criterion tree) moves before the union grows;
the evaluator and policy stay in features/goals.
