+++ lib/classes/goal_criterion.freezed.dart
