+++ lib/classes/goal_enums.dart
