+++ lib/classes/goal_window.dart
