+++ lib/classes/goal_window.freezed.dart
