+++ lib/features/goals/evaluation/goal_progress_evaluator.dart
@@ -1,10 +1,10 @@
 import 'dart:math' as math;
 
+import 'package:lotti/classes/goal_criterion.dart';
+import 'package:lotti/classes/goal_enums.dart';
+import 'package:lotti/classes/goal_window.dart';
 import 'package:lotti/features/goals/evaluation/goal_evaluation.dart';
 import 'package:lotti/features/goals/evaluation/goal_signal_window.dart';
-import 'package:lotti/features/goals/model/goal_criterion.dart';
-import 'package:lotti/features/goals/model/goal_enums.dart';
-import 'package:lotti/features/goals/model/goal_window.dart';
 
 /// Pure fold of a [GoalCriterion] tree over a [GoalSignalWindow].
 ///
