+++ test/features/agents/eval/goal/goal_agent_eval_runner_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:lotti/classes/goal_enums.dart';
 import 'package:lotti/features/ai/model/ai_config.dart';
-import 'package:lotti/features/goals/model/goal_enums.dart';
 
 import '../../../ai_consumption/test_utils.dart';
 import 'support/goal_agent_eval_fixtures.dart';
