+++ test/features/goals/model/goal_criterion_test.dart
@@ -2,9 +2,9 @@ import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/classes/entity_definitions.dart';
-import 'package:lotti/features/goals/model/goal_criterion.dart';
-import 'package:lotti/features/goals/model/goal_enums.dart';
-import 'package:lotti/features/goals/model/goal_window.dart';
+import 'package:lotti/classes/goal_criterion.dart';
+import 'package:lotti/classes/goal_enums.dart';
+import 'package:lotti/classes/goal_window.dart';
 
 void main() {
   group('fromAutoCompleteRule thresholds', () {
